style(workflow): remove unused functions and add function comments
