+++ internal/context/workspace/domain/workflow/event_handlers.go
@@ -1,16 +1,11 @@
 package workflow
 
 import (
-	"bufio"
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/Bio-OS/bioos/pkg/utils"
 	"os"
 	"path"
-	"regexp"
-	"sort"
-	"strings"
 	"time"
 
 	"k8s.io/apimachinery/pkg/util/sets"
@@ -22,7 +17,6 @@ import (
 	apperrors "github.com/Bio-OS/bioos/pkg/errors"
 	applog "github.com/Bio-OS/bioos/pkg/log"
 	"github.com/Bio-OS/bioos/pkg/schema"
-	"github.com/Bio-OS/bioos/pkg/utils/exec"
 	"github.com/Bio-OS/bioos/pkg/utils/git"
 	"github.com/Bio-OS/bioos/pkg/validator"
 )
@@ -144,92 +138,6 @@ func (h *WorkflowVersionAddedHandler) handle(ctx context.Context, workflowID str
 	}
 	version.Graph = graph
 	return nil
-}
-
-func (h *WorkflowVersionAddedHandler) parseWorkflowVersion(_ context.Context, mainWorkflowPath string) (string, error) {
-	versionRegexp := regexp.MustCompile(VersionRegexpStr)
-	file, err := os.Open(mainWorkflowPath)
-	if err != nil {
-		applog.Errorw("fail to open main workflow file", "err", err)
-		return "", err
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		matched := versionRegexp.FindStringSubmatch(line)
-		if matched != nil && len(matched) >= 2 {
-			applog.Infow("version regexp matched", "matched", matched)
-			return matched[1], nil
-		}
-	}
-
-	if err := scanner.Err(); err != nil {
-		return "", err
-	}
-	return "draft-2", nil
-}
-
-func (h *WorkflowVersionAddedHandler) getWorkflowParams(ctx context.Context, name string, arg ...string) ([]WorkflowParam, error) {
-	params := make([]WorkflowParam, 0)
-	outputsResult, err := exec.Exec(ctx, CommandExecuteTimeout, name, arg...)
-	if err != nil {
-		return params, err
-	}
-	var outputsMap map[string]string
-	if err := json.Unmarshal(outputsResult, &outputsMap); err != nil {
-		return params, err
-	}
-
-	for paramName, value := range outputsMap {
-		paramType, optional, defaultValue := parseWorkflowParamValue(value)
-		param := WorkflowParam{
-			Name:     paramName,
-			Type:     paramType,
-			Optional: optional,
-		}
-		if defaultValue != nil {
-			param.Default = *defaultValue
-		}
-		params = append(params, param)
-	}
-	// keep the return sort stable
-	sort.Slice(params, func(i, j int) bool {
-		return params[i].Name < params[j].Name
-	})
-	return params, nil
-}
-
-func parseWorkflowParamValue(value string) (paramType string, optional bool, defaultValue *string) {
-	splitByLeftBracket := strings.SplitN(value, "(", 2)
-	paramType = strings.TrimSpace(splitByLeftBracket[0])
-	if len(splitByLeftBracket) == 1 {
-		return paramType, false, nil
-	}
-
-	extraInfo := strings.TrimSuffix(splitByLeftBracket[1], ")")
-	splitByComma := strings.SplitN(extraInfo, ",", 2)
-	if strings.TrimSpace(splitByComma[0]) == "optional" {
-		optional = true
-	}
-	if len(splitByComma) == 1 {
-		return paramType, optional, nil
-	}
-
-	defaultInfo := strings.TrimSpace(splitByComma[1])
-	splitByEqual := strings.SplitN(defaultInfo, "=", 2)
-	if len(splitByEqual) != 2 || strings.ToLower(strings.TrimSpace(splitByEqual[0])) != "default" {
-		return paramType, optional, nil
-	}
-	rawDefaultValue := strings.TrimSpace(splitByEqual[1])
-	defaultValue = new(string)
-	if strings.HasPrefix(rawDefaultValue, `"`) && strings.HasSuffix(rawDefaultValue, `"`) { // String type
-		_ = json.Unmarshal([]byte(rawDefaultValue), defaultValue) // escape, never error
-	} else {
-		*defaultValue = rawDefaultValue
-	}
-	return paramType, optional, defaultValue
 }
 
 type WorkspaceDeletedHandler struct {
